refactor: moved useApplicationLinksGivenAddress into hooks/applinks
